+++ packages/create-block/CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Master
 
+### Enhancements
+
+- Removed the code that clears the terminal while the block is scaffolded ([#19867](https://github.com/WordPress/gutenberg/pull/19867)).
+
+### Bug Fixes
+
+- Use the description provided to fill the `description` field in `package.json` file in ESNext template ([#19867](https://github.com/WordPress/gutenberg/pull/19867)).
+- Ensure that values provided for slug and namespace get converted to lower case ([#19867](https://github.com/WordPress/gutenberg/pull/19867)).
+
 ### Internal
 
-- Imported from `create-wordpress-block` npm packgage.
+- Imported from `create-wordpress-block` npm packgage ([#19773](https://github.com/WordPress/gutenberg/pull/19773)).